Added ARGB8888 colors support. Fixed z-buffer for real. Camera controls are better now, kinda sucks even now.

+++ inc/Managers/Buffer.inl
@@ -1,7 +1,9 @@
 #ifndef __BUFFER_INL__
 #define __BUFFER_INL__
 
+#include <cstdint>
 #include <vector>
+#include <stdint.h>
 
 namespace ControlSystem
 {
@@ -11,8 +13,8 @@ namespace ControlSystem
         static std::vector<std::vector<float>> original_buffer; // original frame buffer with no depth
         static std::vector<std::vector<float>> depth_buffer;
 
-        static std::vector<std::vector<size_t>> original_frame_buffer; //filled with single color
-        static std::vector<std::vector<size_t>> frame_buffer;
+        static std::vector<std::vector<uint32_t>> original_frame_buffer; //filled with single color
+        static std::vector<std::vector<uint32_t>> frame_buffer;
     };
 }
 #endif
